refactor: move UserTariffBadge to tariff folder
