+++ packages/ui/src/components/tariff/index.ts
@@ -2,3 +2,4 @@ export * from './card';
 export * from './list';
 export * from './row';
 export * from './list-row';
+export * from './badge';
